Rename package: restpagination
