+++ src/main/java/tech/jhipster/lite/generator/client/restpagination/application/TSRestPaginationApplicationService.java
@@ -1,7 +1,7 @@
-package tech.jhipster.lite.generator.client.rest_pagination.application;
+package tech.jhipster.lite.generator.client.restpagination.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.client.rest_pagination.domain.TSRestPaginationModuleFactory;
+import tech.jhipster.lite.generator.client.restpagination.domain.TSRestPaginationModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
